requirements.txt
openai>=1.30.0
supabase>=2.4.0
streamlit>=1.35.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
email-reply-parser>=0.5.12
pytz>=2024.1
